server/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==26.0.0
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.5.0
style==1.1.0
typing_extensions==4.12.2
update==0.0.1
uritemplate==4.1.1
